+++ packages/keystone/src/admin-ui/components/index.ts
@@ -7,7 +7,10 @@ export { ErrorBoundary, ErrorContainer } from './Errors';
 // ADMIN-UI CUSTOM COMPONENTS
 export { Logo } from './Logo';
 export { Navigation, NavigationContainer, NavItem, ListNavItems, ListNavItem } from './Navigation';
-export type { NavigationProps } from './Navigation';
+
+// co-locating the type with the admin-ui/component for a more a salient mental model.
+// importing this type from @keystone-next/keystone/admin-ui/components is probably intuitive for a user
+export type { NavigationProps } from '@keystone-next/types';
 
 // CUSTOM PAGE BUILDING UTILITIES
 export { PageContainer } from './PageContainer';
