+++ packages/keystone/src/admin-ui/utils/useLazyMetadata.tsx
@@ -1,23 +1,10 @@
 import { GraphQLError } from 'graphql';
+import type { AuthenticatedItem, VisibleLists, CreateViewFieldModes } from '@keystone-next/types';
 import { useMemo } from 'react';
 import { DeepNullable, makeDataGetter } from '@keystone-next/admin-ui-utils';
 import { DocumentNode, useQuery, QueryResult, ServerError, ServerParseError } from '../apollo';
 
-export type AuthenticatedItem =
-  | { state: 'unauthenticated' }
-  | { state: 'authenticated'; label: string; id: string; listKey: string }
-  | { state: 'loading' }
-  | { state: 'error'; error: Error | readonly [GraphQLError, ...GraphQLError[]] };
-
-export type VisibleLists =
-  | { state: 'loaded'; lists: ReadonlySet<string> }
-  | { state: 'loading' }
-  | { state: 'error'; error: Error | readonly [GraphQLError, ...GraphQLError[]] };
-
-export type CreateViewFieldModes =
-  | { state: 'loaded'; lists: Record<string, Record<string, 'edit' | 'hidden'>> }
-  | { state: 'loading' }
-  | { state: 'error'; error: Error | readonly [GraphQLError, ...GraphQLError[]] };
+export type { AuthenticatedItem, VisibleLists, CreateViewFieldModes } from '@keystone-next/types';
 
 export function useLazyMetadata(query: DocumentNode): {
   authenticatedItem: AuthenticatedItem;
