+++ packages/types/package.json
@@ -10,7 +10,6 @@
   "dependencies": {
     "@graphql-ts/schema": "0.1.2",
     "@keystone-next/fields": "^12.0.0",
-    "@keystone-next/keystone": "^22.0.0",
     "apollo-server-types": "^0.9.0",
     "cors": "^2.8.5",
     "decimal.js": "10.3.1",
